提交

+++ workspace/solution.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-    long long a, b;
+    int a, b;
     cin >> a >> b;
     cout << a + b << endl;
     return 0;
